fix(pr-reviewer): move parsePrReviewTrigger out of route file
Next.js Route files only allow specific exports (HTTP methods, dynamic,
revalidate, etc.). Exporting parsePrReviewTrigger from route.ts broke
'next build' with "parsePrReviewTrigger is not a valid Route export
field", failing both Lint/Typecheck/Build CI and the Vercel preview.

Move the helper plus EVIDENCE_SIGNAL constants into a sibling module
under lib/agent-runtime so the Route file only exports POST. Behavior is
unchanged; tests continue to pass via the route's import path.
